% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/increase_ages.R
\name{age_pop}
\alias{age_pop}
\title{Increase all agent ages by one time step.}
\usage{
age_pop(pop)
}
\arguments{
\item{pop}{Population data frame}
}
\value{
Updated pop data frame
}
\description{
Updates the 'age' column of the population data frame by adding 1 to all agent ages. Happy birthday!
}
\keyword{internal}
