% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\docType{data}
\name{CoaleDemenyWestF5}
\alias{CoaleDemenyWestF5}
\title{Coale-Demeny West Female Level 5 mortality regime}
\description{
A mortality table used as the default regime in Persephone simulations.
}
\keyword{datasets}
