% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{compute_siler_risk}
\alias{compute_siler_risk}
\title{Compute Siler hazard for a given age}
\usage{
compute_siler_risk(ages, mortality_regime)
}
\arguments{
\item{ages}{Integer vector, if population is age-structured}

\item{mortality_regime}{Data frame with Siler parameters (a1, b1, a2, a3, b3)}
}
\value{
Numeric hazard value
}
\description{
Compute Siler hazard for a given age
}
\keyword{internal}
