% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_starting_population.R
\name{create_pop}
\alias{create_pop}
\title{Create a starting population or age-cohort of agents}
\usage{
create_pop(
  pop0_size,
  age_structured,
  pop_config,
  r = 0,
  mortality_regime = NULL
)
}
\arguments{
\item{age_structured}{Logical: is this an age-structured population? (if not, it is a single age-cohort)}

\item{pop_config}{List of population parameters}

\item{r}{Numeric, the population growth rate}

\item{mortality_regime}{Data frame with Siler parameters (a1, b1, a2, a3, b3)}

\item{pop_size}{Number of agents in the starting pop}

\item{model_lesions}{Logical. If true, then a column for lesion presence is initialized; no one has lesions at current_time = 0.}

\item{lesion_formation_window}{Vector length 2: c(age at which window opens, age at which window closes)}

\item{gammafrailty_shape}{The alpha/shape parameter in a gamma distribution}

\item{gammafrailty_scale}{The sigma/scale parameter in a gamma distribution. Together these two parameters describe the frailty distribution at birth in the starting population.}
}
\value{
A data frame with columns: agent_id, age, lesion, dead, in_sample
}
\description{
Generates a data frame of n agents with columns to hold specified agent states (age, lesion status, etc.), depending on the arguments passed to the function.
}
\keyword{internal}
