% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{get_default_params}
\alias{get_default_params}
\title{Get default simulation parameters}
\usage{
get_default_params()
}
\value{
A named list of default parameter values
}
\description{
Save this function, without any specified arguments, to an object, and that object will be a list of all the default parameter values for Simulate_Cemetery.
}
