% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lesions.R
\name{form_lesions}
\alias{form_lesions}
\title{Apply lesion formation before or after mortality}
\usage{
form_lesions(
  pop,
  lesion_formation_window,
  lesion_formation_rate = NULL,
  annual_exposure = NULL,
  exposure_causes_hazard = FALSE,
  hazard_is_transient = FALSE,
  exposure_hazard_multiplier = 1
)
}
\arguments{
\item{pop}{Population data frame}

\item{lesion_formation_window}{numeric vector of length = 2. c(Age at which lesions can start forming, Age at which lesions stop forming)}

\item{lesion_formation_rate}{A probability}

\item{annual_exposure}{A proportion. If lesion_formation_rate has a value then annual_exposure should be set to NULL, and vice versa.}

\item{exposure_causes_hazard}{Logical. Does exposure to lesion-causing events affect mortality hazard?}

\item{hazard_is_transient}{Logical. If true, exposure-mediated change in mortality hazard last only for the year of exposure.}

\item{exposure_hazard_multiplier}{Numeric. The amount to multiply an agent's mortality hazard if exposure_causes_hazard == TRUE.}
}
\value{
Updated pop data frame
}
\description{
a single vectorized draw to update each agent's lesion state. Agents who do not have a lesion may acquire one during this action, if they meet the conditional requirements defined by the function arguments. Lesions can either form probabilistically, according to a lesion formation rate, or they can form deterministically, according to the proportion of agents exposed to lesion-causing conditions (annual_exposure). Lesion formation rate will be a better predictor of the number of lesions that form, since annual exposure is spread evenly across agents of all ages, and not all ages are capable of forming lesions (ages of lesion formation are determined by the values of the lesion formation window.)

#' Note: These aren't quite parallel options. if(lesion_formation_rate) is used, then acquired frailty values aren't updated. Right now frailty is only integrated with deterministic lesion formation, not probabilistic lesion formation.
}
\keyword{internal}
