% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility.R
\name{generate_births}
\alias{generate_births}
\title{Generate new agent rows to append to the pop}
\usage{
generate_births(pop, current_time, n_births, pop_config)
}
\arguments{
\item{pop}{Population data frame (used to determine next agent_id)}

\item{current_time}{Numeric. Time counter for ABM main loop.}

\item{n_births}{Integer. Number of new agents to create. generate_births() is called inside apply_fertility(), and n_births is calculated inside apply_fertility() before this call, using the birth probabilities calculated from compute_asfr_trapezoid(), should be supplied to apply_fertility() in the asfr argument.}

\item{pop_config}{A list of parameters for the initial population so that new agents will have matching trait columns. Defaults to the pop_config object that should be automatically defined earlier in Simulate_Cemetery, before the main ABM loop.}
}
\value{
Data frame with n_births rows ready to rbind() onto the population
}
\description{
Creates n_births new agents with age 0 and no lesion, with agent_ids
continuing sequentially from the current maximum in the population. All other
columns are initialized to match the structure of create_pop().
This is an internal helper function called by apply_fertility.
}
\keyword{internal}
