% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertility.R, R/simulate.R
\name{generate_births}
\alias{generate_births}
\title{Generate new agent rows to append to the cohort}
\usage{
generate_births(pop, n_births, pop_config)

generate_births(pop, n_births, pop_config)
}
\arguments{
\item{pop}{Population data frame (used to determine next agent_id)}

\item{n_births}{Integer. Number of new agents to create.}

\item{pop_config}{A list of parameters for the initial population so that new agents will have matching trait columns.}

\item{cohort}{Population data frame (used to determine next agent_id)}
}
\value{
Data frame with n_births rows ready to rbind() onto the cohort

Data frame with n_births rows ready to rbind() onto the population
}
\description{
Creates n_births new agents with age 0 and no lesion, with agent_ids
continuing sequentially from the current maximum in the cohort. All other
columns are initialized to match the structure of create_cohort().

Creates n_births new agents with age 0 and no lesion, with agent_ids
continuing sequentially from the current maximum in the population. All other
columns are initialized to match the structure of create_pop().
}
\keyword{internal}
