% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{record_cohort_survivors}
\alias{record_cohort_survivors}
\title{Record survivor snapshot for the current timestep.}
\usage{
record_cohort_survivors(pop, current_time, model_lesions)
}
\arguments{
\item{pop}{Population data frame}

\item{current_time}{Current timestep}

\item{model_lesions}{Logical}
}
\value{
One-row data frame with Age, Alive, Lesion, Lesion_perc
}
\description{
Record survivor snapshot for the current timestep.
}
\keyword{internal}
